backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
sqlalchemy
alembic
loguru
python-multipart
pyyaml